Consolidate File Processing (#1449)
