+++ backend/danswer/connectors/loopio/connector.py
@@ -9,10 +9,6 @@ from requests_oauthlib import OAuth2Session  # type: ignore
 
 from danswer.configs.app_configs import INDEX_BATCH_SIZE
 from danswer.configs.constants import DocumentSource
-from danswer.connectors.cross_connector_utils.html_utils import parse_html_page_basic
-from danswer.connectors.cross_connector_utils.html_utils import (
-    strip_excessive_newlines_and_spaces,
-)
 from danswer.connectors.cross_connector_utils.miscellaneous_utils import time_str_to_utc
 from danswer.connectors.interfaces import GenerateDocumentsOutput
 from danswer.connectors.interfaces import LoadConnector
@@ -22,6 +18,8 @@ from danswer.connectors.models import BasicExpertInfo
 from danswer.connectors.models import ConnectorMissingCredentialError
 from danswer.connectors.models import Document
 from danswer.connectors.models import Section
+from danswer.file_processing.html_utils import parse_html_page_basic
+from danswer.file_processing.html_utils import strip_excessive_newlines_and_spaces
 from danswer.utils.logger import setup_logger
 
 LOOPIO_API_BASE = "https://api.loopio.com/"
