+++ backend/danswer/dynamic_configs/factory.py
@@ -1,4 +1,3 @@
-from danswer.configs.app_configs import DYNAMIC_CONFIG_DIR_PATH
 from danswer.configs.app_configs import DYNAMIC_CONFIG_STORE
 from danswer.dynamic_configs.interface import DynamicConfigStore
 from danswer.dynamic_configs.store import FileSystemBackedDynamicConfigStore
@@ -8,7 +7,7 @@ from danswer.dynamic_configs.store import PostgresBackedDynamicConfigStore
 def get_dynamic_config_store() -> DynamicConfigStore:
     dynamic_config_store_type = DYNAMIC_CONFIG_STORE
     if dynamic_config_store_type == FileSystemBackedDynamicConfigStore.__name__:
-        return FileSystemBackedDynamicConfigStore(DYNAMIC_CONFIG_DIR_PATH)
+        raise NotImplementedError("File based config store no longer supported")
     if dynamic_config_store_type == PostgresBackedDynamicConfigStore.__name__:
         return PostgresBackedDynamicConfigStore()
