+++ backend/requirements/default.txt
@@ -30,7 +30,6 @@ llama-index==0.9.45
 Mako==1.2.4
 msal==1.26.0
 nltk==3.8.1
-docx2txt==0.8
 Office365-REST-Python-Client==2.5.4
 oauthlib==3.2.2
 openai==1.3.5
